+++ tests/test_bigtable/test_parallel_read_rows.py
@@ -21,9 +21,9 @@ import os
 from re import escape
 from .bigtable_emulator import BigtableEmulator
 from tensorflow_io.python.ops import core_ops
-from tensorflow_io.python.ops.bigtable.bigtable_dataset_ops import BigtableClient
-import tensorflow_io.python.ops.bigtable.bigtable_row_range as row_range
-import tensorflow_io.python.ops.bigtable.bigtable_row_set as row_set
+from tensorflow_io.python.ops.bigtable_dataset_ops import BigtableClient
+import tensorflow_io.python.ops.bigtable_row_range as row_range
+import tensorflow_io.python.ops.bigtable_row_set as row_set
 import tensorflow as tf
 from tensorflow import test
